Use copyfile instead of copy2.
This throws away file metadata, but the ownership metadata is
meaningless across a filesystem boundary anyway. It requires less
permission on the parent directory than `copy2`.

Keep in mind that this only affects the `--copy-files` convenience
pathway. The default pathway leaves it to the user to copy the files
with reference to a manifest, and they can use any copy method they
like.